move files into viz-theme
