+++ blog/gatsby-config.js
@@ -2,6 +2,14 @@ module.exports = {
   plugins: [
     `gatsby-theme-rpsych`,
     {
+      resolve: `gatsby-source-filesystem`,
+      options: {
+        path: `content/blog`,
+        name: `blog`,
+        ignore: [`**/**.knit.md`, `**/**.utf8.md`, `**/**.Rmd`, `**/**.R`, `**/*_cache/**`, `**/cache/**`]
+      },
+    },
+    {
       resolve: `gatsby-plugin-manifest`,
       options: {
         name: `R Psychologist - Kristoffer Magnusson`,
