+++ gatsby-theme-rpsych-viz/gatsby-browser.js
@@ -1,3 +1,2 @@
 import "./src/styles/App.css"
-import "katex/dist/katex.min.css"
 import "prismjs/themes/prism-solarizedlight.css"
