+++ gatsby-theme-rpsych-viz/src/components/Footer.js
