+++ gatsby-theme-rpsych-viz/src/components/HeaderAppBar.js
