+++ gatsby-theme-rpsych-viz/src/components/ResponsiveChart.js
