+++ package.json
@@ -13,7 +13,8 @@
     "workspaces": [
       "blog",
       "rpsych-cohend",
-      "gatsby-theme-rpsych"
+      "gatsby-theme-rpsych",
+      "gatsby-theme-rpsych-viz"
     ]
   }
