+++ rpsych-cohend/package.json
@@ -3,45 +3,21 @@
   "version": "2.3.0",
   "private": true,
   "dependencies": {
-    "@material-ui/core": "^4.7.0",
-    "@material-ui/icons": "^4.5.1",
-    "@material-ui/styles": "^4.6.0",
-    "@mdx-js/mdx": "^1.5.8",
-    "@mdx-js/react": "^1.5.8",
-    "@vx/axis": "^0.0.199",
-    "d3": "^5.14.2",
+    "gatsby-theme-rpsych": "*",
+    "gatsby-theme-rpsych-viz": "*"
+  },
+  "peerDependencies": {
     "gatsby": "^2.18.2",
-    "gatsby-image": "^2.2.36",
-    "gatsby-plugin-google-analytics": "^2.1.31",
-    "gatsby-plugin-manifest": "^2.2.33",
-    "gatsby-plugin-material-ui": "^2.1.6",
-    "gatsby-plugin-mdx": "^1.1.3",
-    "gatsby-plugin-offline": "^3.0.29",
-    "gatsby-plugin-react-helmet": "^3.1.18",
-    "gatsby-plugin-sharp": "^2.3.7",
-    "gatsby-plugin-zopfli": "^1.3.2",
-    "gatsby-remark-images": "^3.2.1",
-    "gatsby-remark-katex": "^3.1.17",
-    "gatsby-remark-prismjs": "^3.3.25",
-    "gatsby-source-filesystem": "^2.1.38",
-    "gatsby-transformer-remark": "^2.6.37",
-    "gatsby-transformer-sharp": "^2.3.9",
-    "gatsby-transformer-yaml": "^2.2.20",
-    "jstat": "^1.9.2",
-    "katex": "^0.12.0",
-    "lodash": "^4.17.15",
-    "prismjs": "^1.17.1",
-    "prop-types": "^15.7.2",
     "react": "^16.12.0",
-    "react-color": "^2.18.0",
-    "react-dom": "^16.12.0",
-    "react-helmet": "^5.2.1",
-    "react-joyride": "^2.2.1",
-    "react-scripts": "3.2.0",
-    "react-spring": "^9.0.0-beta.34",
-    "react-use-gesture": "^7.0.5",
-    "svgsaver": "^0.9.0",
-    "typescript": "^3.7.2"
+    "react-dom": "^16.12.0"
+  },
+  "devDependencies": {
+    "eslint": "^6.2.2",
+    "eslint-plugin-react": "^7.11.1",
+    "gatsby": "^2.18.2",
+    "prettier": "^1.14.2",
+    "react": "^16.12.0",
+    "react-dom": "^16.12.0"
   },
   "scripts": {
     "build-production": "gatsby build --prefix-paths",
@@ -55,10 +31,6 @@
   },
   "eslintConfig": {
     "extends": "react-app"
-  },
-  "devDependencies": {
-    "eslint": "^6.7.0",
-    "eslint-plugin-react": "^7.16.0"
   },
   "resolutions": {
     "webpack": "~4.43.0"
