+++ rpsych-cohend/src/Viz.js
@@ -8,7 +8,7 @@ import Box from "@material-ui/core/Box";
 import Cohend from "./components/viz/Overlap";
 import DonutChart from "./components/viz/Donuts";
 import ResponsiveChart from "./components/viz/ResponsiveChart";
-import Slider from "./components/navigation/SettingsSlider";
+import Slider from "./components/settings/SettingsSlider";
 import CircularProgress from '@material-ui/core/CircularProgress';
