+++ rpsych-cohend/src/components/settings/SettingsSlider.js
