+++ rpsych-cohend/src/pages/cohend.js
@@ -1,5 +1,4 @@
 import React from 'react';
 import App from '../App';
-import Typography from '@material-ui/core/Typography';
 
 export default () => <App />
